+++ src/NET/MvvmDialogs/IDialogService.cs
@@ -32,6 +32,22 @@ namespace MvvmDialogs
             where T : Window;
 
         /// <summary>
+        /// Displays a non-modal custom dialog of specified type <typeparamref name="T"/>.
+        /// </summary>
+        /// <param name="ownerViewModel">
+        /// A view model that represents the owner window of the custom dialog.
+        /// </param>
+        /// <param name="viewModel">The view model of the new custom dialog.</param>
+        /// <typeparam name="T">The type of the custom dialog to show.</typeparam>
+        /// <exception cref="ViewNotRegisteredException">
+        /// No view is registered with specified owner view model as data context.
+        /// </exception>
+        void ShowCustom<T>(
+            INotifyPropertyChanged ownerViewModel,
+            INotifyPropertyChanged viewModel)
+            where T : IWindow;
+
+        /// <summary>
         /// Displays a non-modal dialog of a type that is determined by the dialog type locator.
         /// </summary>
         /// <param name="ownerViewModel">
@@ -64,6 +80,26 @@ namespace MvvmDialogs
             INotifyPropertyChanged ownerViewModel,
             IModalDialogViewModel viewModel)
             where T : Window;
+
+        /// <summary>
+        /// Displays a custom modal dialog of specified type <typeparamref name="T"/>.
+        /// </summary>
+        /// <param name="ownerViewModel">
+        /// A view model that represents the owner window of the custom dialog.
+        /// </param>
+        /// <param name="viewModel">The view model of the new custom dialog.</param>
+        /// <typeparam name="T">The type of the custom dialog to show.</typeparam>
+        /// <returns>
+        /// A nullable value of type <see cref="bool"/> that signifies how a window was closed by
+        /// the user.
+        /// </returns>
+        /// <exception cref="ViewNotRegisteredException">
+        /// No view is registered with specified owner view model as data context.
+        /// </exception>
+        bool? ShowCustomDialog<T>(
+            INotifyPropertyChanged ownerViewModel,
+            IModalDialogViewModel viewModel)
+            where T : IWindow;
 
         /// <summary>
         /// Displays a modal dialog of a type that is determined by the dialog type locator.