+++ src/Universal/Demo/Demo.CustomContentDialog/AddTextContentDialog.xaml.cs
@@ -1,0 +1,10 @@
+﻿namespace Demo.CustomContentDialog
+{
+    public sealed partial class AddTextContentDialog
+    {
+        public AddTextContentDialog()
+        {
+            InitializeComponent();
+        }
+    }
+}