+++ src/Universal/Demo/Demo.CustomContentDialog/ViewModelLocator.cs
@@ -1,0 +1,22 @@
+﻿using GalaSoft.MvvmLight.Ioc;
+using Microsoft.Practices.ServiceLocation;
+
+namespace Demo.CustomContentDialog
+{
+    /// <summary>
+    /// This class contains static references to all the view models in the
+    /// application and provides an entry point for the bindings.
+    /// </summary>
+    public class ViewModelLocator
+    {
+        public ViewModelLocator()
+        {
+            ServiceLocator.SetLocatorProvider(() => SimpleIoc.Default);
+
+            SimpleIoc.Default.Register<MainPageViewModel>();
+            SimpleIoc.Default.Register<AddTextCustomContentDialogViewModel>();
+        }
+
+        public MainPageViewModel MainPage => ServiceLocator.Current.GetInstance<MainPageViewModel>();
+    }
+}