+++ src/Universal/MvvmDialogs/DialogService.cs
@@ -109,6 +109,25 @@ namespace MvvmDialogs
         }
 
         /// <summary>
+        /// Begins an asynchronous operation to show the custom <see cref="IContentDialog" /> of
+        /// type <typeparamref name="T" />.
+        /// </summary>
+        /// <typeparam name="T">The type of the custom content dialog to show.</typeparam>
+        /// <param name="viewModel">The view model of the new custom content dialog.</param>
+        /// <returns>
+        /// An asynchronous operation showing the custom dialog. When complete, returns a
+        /// <see cref="ContentDialogResult" />.
+        /// </returns>
+        public IAsyncOperation<ContentDialogResult> ShowCustomContentDialogAsync<T>(INotifyPropertyChanged viewModel)
+            where T : IContentDialog
+        {
+            if (viewModel == null)
+                throw new ArgumentNullException(nameof(viewModel));
+
+            return ShowContentDialogAsync(viewModel, typeof(T));
+        }
+
+        /// <summary>
         /// Begins an asynchronous operation to show the <see cref="ContentDialog" /> of a type that
         /// is determined by the dialog type locator.
         /// </summary>
@@ -265,11 +284,11 @@ namespace MvvmDialogs
         {
             Logger.Write($"Content dialog: {contentDialogType}; View model: {viewModel.GetType()}");
 
-            ContentDialog dialog = CreateContentDialog(contentDialogType, viewModel);
+            IContentDialog dialog = CreateContentDialog(contentDialogType, viewModel);
             return dialog.ShowAsync();
         }
 
-        private ContentDialog CreateContentDialog(
+        private IContentDialog CreateContentDialog(
             Type dialogType,
             INotifyPropertyChanged viewModel)
         {